Trekk ut VelgBarn- og OmBarna-oppsummering i egne filer
